Projectiles can now hit entities which kill them + add hit sfx + remove old debug loggings
A projectile checks whether it's bounding box intersects with the bounding box of other entities. If that is the case, it removes itself and calls hurt() on the entity that was hit by the projectile. At the moment hurt() causes the entity to disappear but obviously this should become a bit more complex soon.

Also add support for debug drawing of the bounding boxes (window.DEBUG_PROJECTILES).